Adding rocBLAS' batched vectors, updating scal_batched tests.
